pyproject.toml
[project]
name = "nl2sql-agent"
version = "0.0.1"
description = "NL2SQL agent backed by a BigQuery semantic layer (scaffold)"
requires-python = ">=3.11"
authors = [{ name = "Your Team" }]
readme = "README.md"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "google-cloud-bigquery>=3.25",
  "google-cloud-logging>=3.10",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.6",
  "mypy>=1.11",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

